+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 0.9.0
 + GH-86: Implement `filterInstanceOf` to filter a stream by type more easily (thanks @nipafx) 
 + Implement `windowed` to provide more options to windowing functions, namely - ability to specify size, how many to skip each time, and whether to include partial windows
++ GH-88: Implement `increasing`, `decreasing`, `nonIncreasing`, and `nonDecreasing` with both `Comparable` stream inputs or using an explicit `Comparator` to appropriately group elements in the input stream to lists in the output stream (thanks @nipafx)
++ Implement `ensureIncreasing`, `ensureDecreasing`, `ensureNonIncreasing` and `ensureNonDecreasing` with both `Comparable` stream inputs or using an explicit `Comparator` to ensure the given stream meets the critiera, or fail exceptionally otherwise
 
 ### 0.8.0 
 + Add support for `orElse()` and `orElseEmpty()` on size-based gatherers to provide a non-exceptional output stream